+++ src/main/java/redis/clients/johm/collections/RedisMap.java
@@ -170,7 +170,7 @@ public class RedisMap<K, V> implements Map<K, V> {
     @Override
     public int size() {
         int repoSize = nest.cat(JOhmUtils.getId(owner)).cat(field.getName())
-                .hlen();
+                .hlen().intValue();
         return repoSize;
     }
